Update README.md
This script assumes you already have a working installation of the bitcoind server installed and running on the server. This has only been tested on a server running Ubuntu 12.04.

You will also need to install Apache, PHP and MySql on the server. You will also need to install php-curl. There is one database and three tables that will need to be installed and the schema for them is included. One table (named watching) is a temp table used while parsing a transaction. Once a transaction has completed being processed it will move the results from that temp table into the others for storage and as a permanent record. 

It will also connect to a remote server on, let's say, a shared hosting server and will send the results there via curl.